(PC-16837) ComeBackLater: move this page to /identificationStart folder
